the my flashcards page now renders buttons dynamically for each title collected from the database
