+++ src/components/FlashCards.tsx
@@ -1,16 +1,16 @@
+import { useLocation } from "react-router-dom";
+
 function FlashCards(){
 
-    interface CardTitles {
-        title: string
-    }
+    const fetchedFolders = useLocation();
 
-    const folders:string[] = [];
+    // fetchedFolders.state
 
-    fetch("http://localhost:8080/getCards")
-    .then((res) => res.json())
-    .then((data) => {data.forEach((e: CardTitles)=>
-      
-        folders.push((e.title))); console.log(data)});
+    const teststring = [];
+    for (let i=0; i < fetchedFolders.state.length; i++){
+  
+      teststring.push(<button key={`${[i]}`} onClick={()=>{console.log()}} className='displayCards'>{fetchedFolders.state[i]}</button>)}
+
   
     return (
         <>
@@ -18,26 +18,10 @@ function FlashCards(){
             
             <h2 className='PageTitle'>Flashcards</h2>
 
-            <div id='myFlashCards'className='childDiv'>
+            <div className='myFlashCards'>
 
-                <button id='button1' className='FlashCardsButtons' onClick={()=>{console.log(folders)}}>
-                    <img height='50%' width='50%' src='../src/assets/black-folder.png' alt="a black folder"/>
-                    <p>{folders[0]}</p>
-                </button>
- 
-                <button id='folder2' className='FlashCardsButtons' onClick={()=>{
-
-                for (let i=0; i < folders.length; i++){
-                    const buttonFolders = document.createElement("button")
-                    buttonFolders.textContent=folders[i]
-                    document.getElementById('myFlashCards').appendChild(buttonFolders)
-                }
-                }}>
-                    <img height='50%' width='50%' src='../src/assets/black-folder.png' alt="a black folder"/>
-              
-            </button>
+                {[...teststring]}
             
-
